Break the main function into separate concerns
This will allow for better testing. Before, everything was tightly coupled and
the server would simply be started, and then the results printed to a terminal.
Tests could only start the server as a separate process and parse its stdout to
get the results. This way, the server can be started in a separate thread
internal to the test, and the resulting object can be passed back into the test
for assertions. The test can create a channel object that the server library
will use to send received objects back for the assertions.